Added math tag insertion.
Added the command C-c C-m to insert math tags around the cursor or selected
region. If the cursor is inside a math environment, no math tags will be
added.

This implementation is not complete yet. In the future the command C-c C-m
should open a math environment for editing via AucTeX. This behaviour will be
independent of adding math tags. This means that the command will add math
tags in case there are none, and opens the math editing afterwards.