Disabled automated line breaks.
The auto-fill-mode which was enabled by default inserts linebreaks
automatically. Usually this improves the readability of the text. Since the
MediaWiki code written online has "infinitely long" lines, this induces big
edits, when only a few characters are changed. To still keep the readability
as good as possible the visual-line-mode is enabled instead. This displays the
without breaking single words apart in long lines.