Yasnippets: Add function to setup provided snippets
Store the location of this folder in the varibale
mediawiki-mode-snippets-dir. To ease adding the snippets to yasnippet, the
contains the function mediawiki-mode-snippet-initialize which adds the
snippets provided by this package to yasnippet.

This is an optional addition to mediawiki-mode. Thus the variable and function
are only defined, when yasnippet is installed.